Fix memory leak of ext_host_name
ssl_ext has a hostname which is strdup()'d (i.e. allocated) but wasn't
freed in the ssl_ext_free() routine.  This lead to the leaking of that
strdup()'d pointer on every SSL close/reopen cycle.

Patch simply checks if there's a string and frees it in the ssl_ext_free
routine.  Also adds a sanity check in set_ext_hostname() to not try and
free() the old string if there is not one present.