Fix handling of padded TLS fragments in handshake
When multiple handshake messages are sent in a single fragment, there
may be some padding (i.e. pkt_size will be larger than the amount of bytes
processed).  When this happens, the old code would only advance the working
pointer to the end of processed data, which would not be the start of the
next packet per the sent pkt_size, causing handshake failure.

Now simply advance the working pointer to the next packet irrespective
of how many bytes in the current one were processed in the server_hello
message.

Also fix a CI problem introduced when the Arduino core common.sh started
checking for a valid defined BUILD_TYPE.